MPV/requirements.txt
pyserial==3.5
pillow==10.0.0